+++ src/api/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
   fields validation regex patterns
 - Ensure the `Session.energy` field is in [0 ; 1000]
 - Ensure session do not last more than a week
+- Ensure `Statique.coordonneesXY` is located in France
 
 #### Dependencies